hotfix: useSearchParams <Suspense> 에러 바운더리 처리
